requirements.txt
python-telegram-bot==20.8
google-generativeai==0.7.2
requests==2.31.0